requirements.txt
fastmcp[apps]>=3.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
anthropic>=0.21.0
prefab-ui>=0.1.0
